refactor: svg 컴포넌트 디렉토리에 index.ts 만들어 여러 컴포넌트 동시 import 시 코드 중복 없앰
* npm run svgr 스크립트 실행 시 index.ts 자동 생성되도록 함
