+++ src/components/Navigation/DesktopNavigation.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import TextButton from "@/components/TextButton";
-import SvgDddIcon from "@/components/svgs/DddIcon";
+import { DddIcon } from "@/components/svgs";
 import navigationList from "@/fixtures/navigationList.json";
 
 export default function DesktopNavigation() {
@@ -10,7 +10,7 @@ export default function DesktopNavigation() {
   return (
     <div className="flex w-full justify-center pt-32">
       <div className="bg-black w-[55px] h-[55px] rounded-full flex justify-center items-center flex-shrink-0">
-        <SvgDddIcon />
+        <DddIcon />
       </div>
       <nav className="bg-white p-4 rounded-full flex gap-[2px]">
         {navigationList.map((item) => (
