+++ src/components/TextInput/index.tsx
@@ -1,6 +1,6 @@
 import { cva, type VariantProps } from "class-variance-authority";
 import { cn } from "@/utils/cn";
-import ErrorIcon from "@/components/svgs/ErrorIcon";
+import { ErrorIcon } from "@/components/svgs";
 
 const inputVariants = cva(
   "relative rounded-[99px] border-[1.5px] border-solid border-gray-30 focus:border-blue-40 disabled:border-gray-20 disabled:bg-gray-10 disabled:text-gray-30",
