+++ src/components/svgs/index.ts
@@ -1,0 +1,9 @@
+export { default as AddIcon } from "./AddIcon";
+export { default as CloseIcon } from "./CloseIcon";
+export { default as DddIcon } from "./DddIcon";
+export { default as ErrorIcon } from "./ErrorIcon";
+export { default as GithubIcon } from "./GithubIcon";
+export { default as InstagramIcon } from "./InstagramIcon";
+export { default as LinkedInIcon } from "./LinkedInIcon";
+export { default as MenuIcon } from "./MenuIcon";
+export { default as SubtractIcon } from "./SubtractIcon";
